bug fix implicit conversion

+++ BGWpy/core/F90io.py
@@ -50,7 +50,7 @@ def fortran_str(obj, quotes=True):
         return bool_str(obj)
     if '__iter__' in dir(obj):
         try:
-            return arr_str(array(obj, dtype=double))
+            return arr_str(array(obj)) # Daan ; let numpy figure out whether the object should be a int or float
         except:
             return ' '.join([fortran_str(o, quotes) for o in obj])
